Extract pool bounds aggregation into a shared module
The per-pool bounds trackers each carried a `compute_pool_bounds` method
holding the pool-specific aggregation. Pull those into a shared `pool_bounds`
module as free functions `compute_pv_pool_bounds` and
`compute_battery_pool_bounds` (with their tests); the trackers now call them.
This isolates the aggregation so the tracker loop can be made generic next.
