+++ src/microgrid.rs
@@ -4,6 +4,7 @@
 //! High-level interface for the Microgrid API.
 
 mod bounds_aggregation;
+mod pool_bounds;
 
 mod battery_bounds_tracker;
 mod battery_pool;
